server: refactor auth package to being package-centric
Addresses: #45
